clean BathToken deprecated logic

+++ contracts/rubiconPoolsv0/BathToken.sol
@@ -165,20 +165,21 @@ contract BathToken is IBathToken {
         underlyingToken.transfer(msg.sender, r);
     }
 
-    function rebalance(address sisterBath, address underlying /* sister asset */, uint stratProportion)
-        external
-        onlyPair
-    {   
+    // This function returns filled orders to the correct liquidity pool and sends strategist rewards to the Pair
+    function rebalance(
+        address sisterBath,
+        address underlying, /* sister asset */
+        uint256 stratProportion
+    ) external onlyPair {
         require(stratProportion > 0 && stratProportion < 20);
-        uint stratReward = stratProportion * (IERC20(underlying).balanceOf(address(this))) / 100;
+        uint256 stratReward =
+            (stratProportion * (IERC20(underlying).balanceOf(address(this)))) /
+                100;
         IERC20(underlying).transfer(
             sisterBath,
             IERC20(underlying).balanceOf(address(this)) - stratReward
         );
-        IERC20(underlying).transfer(
-            msg.sender,
-            stratReward
-        );
+        IERC20(underlying).transfer(msg.sender, stratReward);
     }
 
     // *** Internal Functions ***
